Booking Service
Added the Booking service and realized I forgot the DTO for the AdSpace controller.

+++ backend/src/main/java/org/example/backend/Repository/AdSpaceRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 
 @Repository
-public interface AdSpaceRepository extends JpaRepository<AdSpace,Long> {
+public interface AdSpaceRepository extends JpaRepository<AdSpace,Integer> {
     Optional<AdSpace> findById(Long id);
 
 }
